+++ CHANGES.markdown
@@ -1,3 +1,7 @@
 Oct 1, 2012:
 
 - Merged in austinrfnd's `author` handling code.
+
+Dec 13, 2012
+
+- Use fastimage instead of mini_magick for JRuby compat.